Fixed overflow bug per Walter, removed back_search_extend, added a draft of comp-to-merge_P code in comp_P.

+++ frame_2D_alg/frame_blobs.py
@@ -105,6 +105,8 @@ def comp_pixel(image):  # 2x2 pixel cross-correlation within image, a standard e
     # see comp_pixel_versions file for other versions and more explanation
 
     # input slices into sliding 2x2 kernel, each slice is a shifted 2D frame of grey-scale pixels:
+    image = image.astype('int32')  # it appears that we had overflow in G computation
+
     topleft__ = image[:-1, :-1]
     topright__ = image[:-1, 1:]
     bottomleft__ = image[1:, :-1]
@@ -113,7 +115,7 @@ def comp_pixel(image):  # 2x2 pixel cross-correlation within image, a standard e
     rot_Gy__ = bottomright__ - topleft__  # rotated to bottom__ - top__
     rot_Gx__ = topright__ - bottomleft__  # rotated to right__ - left__
 
-    G__ = (np.hypot(rot_Gy__, rot_Gx__) - ave).astype('int')
+    G__ = (np.hypot(rot_Gy__, rot_Gx__) - ave).astype('int32')
     # deviation of central gradient per kernel, between four vertex pixels
     M__ = int(ave * 1.2) - (abs(rot_Gy__) + abs(rot_Gx__))
     # inverse deviation of SAD, which is a measure of variation. Ave * coeff = ave_SAD / ave_G, 1.2 is a guess
